pass concrete argument types when streaming signature as functionInfo

+++ server/src/test/java/io/crate/execution/dsl/projection/WindowAggProjectionSerialisationTest.java
@@ -21,6 +21,20 @@
 
 package io.crate.execution.dsl.projection;
 
+import static java.util.Collections.singletonList;
+import static org.hamcrest.CoreMatchers.is;
+import static org.hamcrest.collection.IsIterableContainingInOrder.contains;
+import static org.junit.Assert.assertThat;
+
+import java.io.IOException;
+import java.util.List;
+
+import org.elasticsearch.Version;
+import org.elasticsearch.common.inject.ModulesBuilder;
+import org.elasticsearch.common.io.stream.BytesStreamOutput;
+import org.hamcrest.Matchers;
+import org.junit.Test;
+
 import io.crate.analyze.WindowDefinition;
 import io.crate.execution.engine.aggregation.impl.AggregationImplModule;
 import io.crate.execution.engine.aggregation.impl.SumAggregation;
@@ -32,19 +46,6 @@ import io.crate.metadata.FunctionImplementation;
 import io.crate.metadata.Functions;
 import io.crate.metadata.functions.Signature;
 import io.crate.types.DataTypes;
-import org.elasticsearch.Version;
-import org.elasticsearch.common.inject.ModulesBuilder;
-import org.elasticsearch.common.io.stream.BytesStreamOutput;
-import org.hamcrest.Matchers;
-import org.junit.Test;
-
-import java.io.IOException;
-import java.util.List;
-
-import static java.util.Collections.singletonList;
-import static org.hamcrest.CoreMatchers.is;
-import static org.hamcrest.collection.IsIterableContainingInOrder.contains;
-import static org.junit.Assert.assertThat;
 
 public class WindowAggProjectionSerialisationTest {
 
@@ -140,10 +141,10 @@ public class WindowAggProjectionSerialisationTest {
         return functions.getQualified(
             Signature.aggregate(
                 SumAggregation.NAME,
-                DataTypes.FLOAT.getTypeSignature(),
-                DataTypes.FLOAT.getTypeSignature()),
-            List.of(DataTypes.FLOAT),
-            DataTypes.FLOAT
+                DataTypes.LONG.getTypeSignature(),
+                DataTypes.LONG.getTypeSignature()),
+            List.of(DataTypes.LONG),
+            DataTypes.LONG
         );
     }
 }
